Re-encoding
Re-encoded the existing FLAC files to the new OPUS-standard.
In short words: we're keeping the quality as it is, while lowering the file size for each music file.

Also, updated the album.json files so the game will be able to correctly play the music.

Also^2, updated the license.txt files for both albums since it's 2022